+++ scripts/events/midi.js
@@ -337,7 +337,7 @@ async function rollFinished(workflow) {
     await executeMacroPass(workflow, 'rollFinished');
     await executeTargetMacroPass(workflow, 'targetRollFinished');
     await executeTargetMacroPass(workflow, 'onHit', true);
-    await executeMacroPass(workflow, 'rollFinishedLate');
+    //await executeMacroPass(workflow, 'rollFinishedLate');
     let sceneTriggers = [];
     workflow.token?.document.parent.tokens.filter(i => i.uuid !== workflow.token?.document.uuid && i.actor).forEach(j => {
         sceneTriggers.push(...getSortedTriggers({token: j.object, actor: j.actor, sourceToken: workflow.token}, 'sceneRollFinished'));