+++ scripts/macros/2014/classFeatures/illrigger/interdictBoons.js
@@ -1,0 +1,5 @@
+export let interdictBoons = {
+    name: 'Interdict Boons',
+    version: '1.3.76',
+    rules: 'legacy'
+};