+++ scripts/macros/2014/spells/banishment.js
@@ -19,7 +19,7 @@ export async function banishmentHelper(workflow, otherPlanarTokens = []) {
         name: workflow.item.name,
         img: workflow.item.img,
         origin: workflow.item.uuid,
-        duration: itemUtils.convertDuration(workflow.item),
+        duration: itemUtils.convertDuration(workflow.activity),
         changes: [
             {
                 key: 'flags.midi-qol.superSaver.all',