+++ scripts/macros/2024/classFeatures/cleric/light/wardingFlare.js
@@ -4,7 +4,7 @@ async function attack({trigger, workflow}) {
 }
 async function use({trigger, workflow}) {
     let item = itemUtils.getItemByIdentifier(workflow.actor, 'improvedWardingFlare');
-    if (item) await workflowUtils.syntheticItemRoll(item, Array.from(workflow.targets), {consumeResources: true, consumeUsage: true, userId: socketUtils.firstOwner(item)});
+    if (item) await workflowUtils.syntheticItemRoll(item, Array.from(workflow.targets), {consumeResources: true, consumeUsage: true, userId: socketUtils.firstOwner(item.actor)});
 }
 export let wardingFlare = {
     name: 'Warding Flare',